+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.119.0
+
+- `[initsystem]` Added [launchd](https://www.launchd.info) support
+- `[support/services]` Added package for collecting services info
+
 ### 12.118.0
 
 - `[terminal/input]` Added method `SetHistoryCapacity`